requirements.txt
django
django-debug-toolbar

# For creating test data
factory_boy